Extract client part
