+++ config/routes.rb
@@ -1,8 +1,10 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  mount Rswag::Api::Engine => '/api/docs' unless Rails.env.production?
+  mount Rswag::Api::Engine => '/api/client/docs' unless Rails.env.production?
   namespace :api do
-    resources :articles, only: %i[index show create update destroy]
+    namespace :client do
+      resources :articles, only: %i[index show create update destroy]
+    end
   end
 end
